requirements.txt
python-dotenv>=1.0.1
streamlit>=1.36.0
assemblyai>=0.46.0
elevenlabs>=2.24.0
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.5.4
langgraph-cli[inmem]>=0.3.1
langsmith>=0.3.37
langchain-community>=0.3.9
langchain-openai>=0.3.28
langchain-anthropic>=0.3.15
langchain-mcp-adapters>=0.1.6
langchain-deepseek>=0.1.2
langchain-tavily
langchain-groq>=0.2.4
langchain-google-vertexai>=2.0.25
langchain-google-genai>=2.1.5
langchain-aws>=0.2.28
openai>=1.99.2
tavily-python>=0.5.0
arxiv>=2.1.3
pymupdf>=1.25.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
duckduckgo-search>=3.0.0
exa-py>=1.8.8
requests>=2.32.3
beautifulsoup4==4.13.3
pytest
httpx>=0.24.0
markdownify>=0.11.6
azure-identity>=1.21.0
azure-search>=1.0.0b2
azure-search-documents>=11.5.2
rich>=13.0.0
ipykernel>=6.29.5
supabase>=2.15.3
mcp>=1.9.4
pandas>=2.3.1
aiohttp>=3.9.5
pydantic>=2.8.2
typing-extensions>=4.12.2
python-dateutil>=2.9.0.post0
